Collapse "tool-search-internal".
