+++ packages/a2/src/tools/search-internal.ts
@@ -2,91 +2,29 @@
  * @fileoverview Search using internal search engine.
  */
 
-import {
-  Capabilities,
-  LLMContent,
-  Outcome,
-  Schema,
-} from "@breadboard-ai/types";
-import { Template } from "../a2/template";
-import { defaultLLMContent, err, ok, toLLMContent, toText } from "../a2/utils";
-import toolSearchInternal, {
-  describe as toolSearchInternalDescribe,
-} from "./tool-search-internal";
+import { Capabilities, Outcome, Schema } from "@breadboard-ai/types";
+import { ok } from "../a2/utils";
+import { executeTool } from "../a2/step-executor";
 export { invoke as default, describe };
 
-type Inputs =
-  | {
-      context?: LLMContent[];
-      "p-query": LLMContent;
-    }
-  | {
-      query: string;
-    };
+type Inputs = {
+  query: string;
+};
 
-type Outputs =
-  | {
-      context: LLMContent[];
-    }
-  | {
-      results: string;
-    };
-
-async function resolveInput(
-  caps: Capabilities,
-  inputContent: LLMContent
-): Promise<string> {
-  const template = new Template(caps, inputContent);
-  const substituting = await template.substitute({}, async () => "");
-  if (!ok(substituting)) {
-    return substituting.$error;
-  }
-  return toText(substituting);
-}
+type Outputs = {
+  results: string;
+};
 
 async function invoke(
-  inputs: Inputs,
+  { query }: Inputs,
   caps: Capabilities
 ): Promise<Outcome<Outputs>> {
-  console.log("INTERNAL SEARCH INPUTS", inputs);
-  let query: string;
-  let mode: "step" | "tool";
-  if ("context" in inputs) {
-    mode = "step";
-    const last = inputs.context?.at(-1);
-    if (last) {
-      query = toText(last);
-    } else {
-      return err("Please provide a query");
-    }
-  } else if ("p-query" in inputs) {
-    query = await resolveInput(caps, inputs["p-query"]);
-    mode = "step";
-  } else {
-    query = inputs.query;
-    mode = "tool";
-  }
-  query = (query || "").trim();
-  if (!query) {
-    return err("Please provide a query");
-  }
-  console.log("Query: " + query);
-  const searchResults = await toolSearchInternal(
-    {
-      query,
-    },
-    caps
-  );
-  if (!ok(searchResults)) {
-    return searchResults;
-  }
-  const results = searchResults.results;
-  if (mode === "step") {
-    return {
-      context: [toLLMContent(results)],
-    };
-  }
-  return { results };
+  const executing = await executeTool<string>(caps, "enterprise_search", {
+    query,
+  });
+  if (!ok(executing)) return executing;
+
+  return { results: executing };
 }
 
 export type DescribeInputs = {
@@ -95,56 +33,29 @@ export type DescribeInputs = {
   asType?: boolean;
 };
 
-async function describe({ asType: _, ...inputs }: DescribeInputs) {
-  const isTool = inputs && Object.keys(inputs).length === 1;
-  if (isTool) {
-    return toolSearchInternalDescribe();
-  }
-  const hasWires = "context" in (inputs.inputSchema.properties || {});
-  const query: Schema["properties"] = hasWires
-    ? {}
-    : {
-        "p-query": {
-          type: "object",
-          title: "Search query",
-          description: "Please provide a search query",
-          behavior: [
-            "llm-content",
-            "config",
-            "hint-preview",
-            "hint-single-line",
-          ],
-          default: defaultLLMContent(),
-        },
-      };
-
+async function describe() {
   return {
     title: "Internal Search",
     description: "Search using internal search engine.",
     inputSchema: {
       type: "object",
       properties: {
-        context: {
-          type: "array",
-          items: { type: "object", behavior: ["llm-content"] },
-          title: "Context in",
-          behavior: ["main-port"],
+        query: {
+          type: "string",
+          title: "Query",
+          description: "The search query",
         },
-        ...query,
       },
     } satisfies Schema,
     outputSchema: {
       type: "object",
       properties: {
-        context: {
-          type: "array",
-          items: { type: "object", behavior: ["llm-content"] },
-          title: "Context out",
-          behavior: ["main-port"],
+        results: {
+          type: "string",
+          title: "Search Results",
         },
       },
     } satisfies Schema,
-
     metadata: {
       icon: "web-search",
       tags: ["quick-access", "tool", "component", "environment-corp"],
